DROTH-2702 change numeric to bigint

+++ digiroad2-oracle/src/main/resources/db/migration/V1__create_tables.sql
@@ -440,7 +440,7 @@ CREATE TABLE manoeuvre_element_history
 
 CREATE TABLE manoeuvre_exceptions
 (
-    manoeuvre_id   numeric(38) NOT NULL,
+    manoeuvre_id  bigint NOT NULL,
     exception_type smallint    NOT NULL
 );
